Ground Station/requirements.txt
aiosqlite==0.19.0
asgiref==3.7.2
blinker==1.6.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.0
Flask==3.0.0
fonttools==4.43.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.26.0
packaging==23.2
Pillow==10.0.1
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.8.0
Werkzeug==3.0.1
